WIP: Rewrite cmake files to be more generic and structured
This is a work in progress to rewrite the cmake configuration files
to yield a working cmake build system for Windows, Linux and OSX.

Old cmake files contain a lot of platform specifc code that
sometimes is not implemented for all supported platforms.

This is an attempt to clean up the configurations files and
have them share as much code as possible between all platforms.