Fixed haltspp/halttime not properly stopping the renderer when using SPPM.
The issue is caused by a race condition.

While the primary thread checks the halt condition, the other threads race past the pause() call, and starts waiting at the barrier after the pause().

If the primary thread terminates the renderer, it will exit on the following pause() call, thus never reaching the following barrier, leaving the other threads stuck in the barrier.

Adding a barrier before the pause() call ensures that this cannot occur.