An algorithm for determining the optimal large mutation probability value for arbitrary scene.
To appear first in LuxRender. Thank you Omniflux for implementing the metric infrastructure, and Lord Crc for the implementation tips, insights and testing. Details will be covered in the paper and tech report. Please give it lots of testing, and report any issues to zsolnai@iit.bme.hu.

You can enable it by adding the following to your .lxs file under the Sampler "metropolis" block:
	"bool adaptive_largemutationprob" ["true"]