Fixed files being erronously reported as invalid when transfered over network, see issue #1239.
The issue was caused by incorrect usage of AdjustFilename after the introduction of the hashed network file transfer.

The original scene filename was hashed, however that file might not exist thus the hash would equal the "empty" hash.
However AdjustFilename was used in the routine sending the file, so the correct file data was sent, causing the server to compute the correct hash.

The fix is to move AdjustFilename earlier, so that the adjusted filename is added to the CompiledFiles container.