Partially fix dispersion issues for path tracing and bidir
The dispersion flag (stored in sw.single) was global for a path. If the path is
dispersed at the last vertex, all previous contributions (stored in L and V)
where considered as dispersed.

Now the PartialContribution class stores differently the contributions
depending if they are dispersed or not.

This should improves the convergence speed on un-dispersed path which appears as
dispersed.

This fully solves the issue for path tracing, and partially sort it for
bidirectional, see next commit.